+++ resources/views/fields/code.blade.php
@@ -5,7 +5,7 @@
             data-fields--code-line-numbers="{{$lineNumbers}}"
             data-fields--code-default-Theme="{{$defaultTheme}}"
     >
-        <div class="code b pos-rlt w-100" style="min-height: 300px;"></div>
+        <div class="code b pos-rlt w-100" style="min-height: {{ $attributes['height'] }}"></div>
         <input type="hidden" @include('platform::partials.fields.attributes', ['attributes' => $attributes])>
     </div>
 @endcomponent