+++ src/Screen/Fields/TinyMCEField.php
@@ -39,7 +39,7 @@ use Orchid\Screen\Field;
  * @method $this value($value = true)
  * @method $this help($value = true)
  * @method $this popover($value = true)
- * @method $this height($value = true)
+ * @method $this height($value = '300px')
  */
 class TinyMCEField extends Field
 {